Preliminary support for building on Watcom C 2.0.c http://sourceforge.net/projects/openwatcom/?source=typ_redirect
The Causeway DOS Extender supports DLL's so this is our 'way out'.  NULL builds, and loads up but crashes when we go to start the actual map.

I haven't debugged yet.

This fixes some things that the Watcom compiler was complaining about, and I verified I'm still able to build a working game with DJGPP

+++ null/cl_null.c
@@ -20,7 +20,7 @@ void CL_Shutdown (void)
 {
 }
 
-void CL_Frame (int msec)
+void CL_Frame (double msec)
 {
 }
 
@@ -30,10 +30,10 @@ void Con_Print (char *text)
 
 void Cmd_ForwardToServer (void)
 {
-	char *cmd;
+        char *cmd;
 
-	cmd = Cmd_Argv(0);
-	Com_Printf ("Unknown command \"%s\"\n", cmd);
+        cmd = Cmd_Argv(0);
+        Com_Printf ("Unknown command \"%s\"\n", cmd);
 }
 
 void SCR_DebugGraph (float value, int color)
@@ -50,6 +50,7 @@ void SCR_EndLoadingPlaque (void)
 
 void Key_Init (void)
 {
-	Cmd_AddCommand ("bind", Key_Bind_Null_f);
+        Cmd_AddCommand ("bind", Key_Bind_Null_f);
+        //Cmd_AddCommand ("unbindall",Key_Unbindall_f);
 }
